make TrackSection a class and remove TrackSectionC
